refactor: Centralized content-tag code
