+++ packages/ast/template-tag/src/-private/content-tag.ts
@@ -1,0 +1,4 @@
+export * from './content-tag/marker.js';
+export * from './content-tag/preprocessor.js';
+export * from './content-tag/slice-byte-range.js';
+export * from './content-tag/types.js';
