+++ packages/ast/template-tag/src/-private/content-tag/marker.ts
@@ -1,0 +1,1 @@
+export const MARKER = 'template_fd9b2463e5f141cfb5666b64daa1f11a';
