+++ packages/ast/template-tag/src/-private/content-tag/preprocessor.ts
@@ -1,0 +1,3 @@
+import { Preprocessor } from 'content-tag';
+
+export const preprocessor = new Preprocessor();
