+++ packages/ast/template-tag/src/-private/content-tag/types.ts
@@ -1,0 +1,1 @@
+export type { Range, Parsed as TemplateTag } from 'content-tag';
