+++ packages/ast/template-tag/src/replace-template-tag.ts
@@ -1,27 +1,4 @@
-import type { Range } from 'content-tag';
-
-const BufferMap = new Map<string, Buffer>();
-
-function getBuffer(str: string): Buffer {
-  let buffer = BufferMap.get(str);
-
-  if (!buffer) {
-    buffer = Buffer.from(str);
-    BufferMap.set(str, buffer);
-  }
-
-  return buffer;
-}
-
-function sliceByteRange(
-  str: string,
-  indexStart: number,
-  indexEnd?: number,
-): string {
-  const buffer = getBuffer(str);
-
-  return buffer.slice(indexStart, indexEnd).toString();
-}
+import { type Range, sliceByteRange } from './-private/content-tag.js';
 
 /**
  * Replaces a particular `<template>` tag.
